use u32 everywhere in lexer to reduce size of structs

+++ src/lexer/parse_heredoc.rs
@@ -125,7 +125,7 @@ impl<'b> Lexer<'b> {
 
         let heredoc_end: HeredocEnd;
 
-        let eos = self.buffer.input.line_at(here.lastline).start + here.offset;
+        let eos = self.buffer.input.line_at(here.lastline).start as u32 + here.offset;
         let len = here.length;
         let func = here.func;
         let indent = here.func & STR_FUNC_INDENT;
@@ -154,7 +154,7 @@ impl<'b> Lexer<'b> {
 
         if (func & STR_FUNC_EXPAND) == 0 {
             loop {
-                ptr = self.buffer.input.line_at(self.buffer.lastline).start;
+                ptr = self.buffer.input.line_at(self.buffer.lastline).start as u32;
                 ptr_end = self.buffer.pend;
                 if ptr_end > ptr {
                     match self.buffer.input.unchecked_byte_at(ptr_end - 1) {
@@ -302,7 +302,7 @@ impl<'b> Lexer<'b> {
         HeredocEnd { start, end, value }
     }
 
-    fn here_document_error(&mut self, here: &HeredocLiteral, eos: usize, len: usize) -> i32 {
+    fn here_document_error(&mut self, here: &HeredocLiteral, eos: u32, len: u32) -> i32 {
         self.heredoc_restore(here);
         self.compile_error(
             DiagnosticMessage::UnterminatedHeredoc {
@@ -352,8 +352,8 @@ impl<'b> Lexer<'b> {
         self.strterm = None;
         let line = here.lastline;
         self.buffer.lastline = line;
-        self.buffer.pbeg = self.buffer.input.line_at(line).start;
-        self.buffer.pend = self.buffer.pbeg + self.buffer.input.line_at(line).len();
+        self.buffer.pbeg = self.buffer.input.line_at(line).start as u32;
+        self.buffer.pend = self.buffer.pbeg + self.buffer.input.line_at(line).len() as u32;
         self.buffer.pcur = self.buffer.pbeg + here.offset + here.length + here.quote;
         self.buffer.ptok = self.buffer.pbeg + here.offset - here.quote;
         self.buffer.heredoc_end = self.buffer.ruby_sourceline;
